requirements_old.txt
pygame==2.1.3.dev8
Pillow==9.2.0
tkvideo==0.1
opencv-python==4.5.5.62
opencv-contrib-python==4.5.5.64
paho-mqtt==1.6.1
mediapipe==0.9.0.1
numpy==1.21.6
tkintermapview==1.19
geographiclib==2.0
